fix move pages to pages folder
